Added delegate callback
Added a delegate callback to handle selecting a single contact. For now, all this does is add a new cell every time a contact is selected. Eventually, a new cell will be created with the contact selected, and that cell will be expandable to reveal extra information. Eventually, Core Data will filter into this.